Add Sphinx documentation generator
Added support for automatic documentation generation with Sphinx.
Made some basic configurations and set it up in order to work with
Google style docstrings. Also installed the readthedocs theme.
A few more steps might be needed, in order to set it up
the way we like it. At the moment it generates documentation only for
the external modules and not for the ros modules (packages)
(which are deeper in the dir tree).

Signed-off-by: Christos Baziotis <christos.baziotis@gmail.com>